Added ParticipantController tests and added a check for delete if id == null

+++ server/src/main/java/server/api/ParticipantController.java
@@ -50,6 +50,7 @@ public class ParticipantController {
     @Transactional
    @DeleteMapping("/{id}")
    public ResponseEntity deleteParticipant(@PathVariable ("id") UUID id) {
+        if(id==null) return ResponseEntity.badRequest().build();
         if(!repo.existsById(id)) return ResponseEntity.notFound().build();
         repo.deleteById(id);
         return ResponseEntity.ok().build();
